+++ operator/pkg/compare/compare.go
@@ -184,8 +184,8 @@ func genYamlIgnoreOpt(yamlStr string) (cmp.Option, error) {
 	}
 	return cmp.FilterPath(func(curPath cmp.Path) bool {
 		up := pathToStringList(curPath)
-		treeNode, found, _ := tpath.GetFromTreePath(tree, up)
-		return found && tpath.IsLeafNode(treeNode)
+		treeNode, found, _ := tpath.Find(tree, up)
+		return found && IsLeafNode(treeNode)
 	}, cmp.Ignore()), nil
 }
 
@@ -496,4 +496,9 @@ func writeStringSafe(sb io.StringWriter, s string) {
 	if err != nil {
 		log.Error(err.Error())
 	}
+}
+
+// IsLeafNode reports whether the given node is a leaf, assuming internal nodes can only be maps or slices.
+func IsLeafNode(node interface{}) bool {
+	return !util.IsMap(node) && !util.IsSlice(node)
 }
